% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_qa_data.R
\name{prepare_qa_data}
\alias{prepare_qa_data}
\title{prepare_qa_data}
\usage{
prepare_qa_data(activityInfoTable)
}
\arguments{
\item{activityInfoTable}{activityInfo link to the db where indicators got reported}
}
\value{
frame with the content of the activity info DB plus indicator reference
        and applied QA
}
\description{
Note that this function is specific to a specific DB schema -
the form to query as well as the columns variable name are hard coded and 
would need to be adjusted if the DB change
}
\details{
Implemented QA automatic Check
 Missing M&E Data sources
 Missing M&E Activity
 Missing Baseline
 Missing Baseline Numerator
 Missing Baseline Denominator
 Missing Actual
 Missing Actual Numerator
 Missing Actual Denominator
 Incorrect Baseline (i.e. >100%)
 Incorrect Actual (i.e. >100%)
 Baseline Higher (Lower) than Actual
 Significant Variance between Baseline and Actual (+/- 2*Baseline)
 Missing Data Limitation for Baseline
 Missing Data Limitation for Actual


Note also that the function is based on a token stored as an environment 
variable

 Credentials set up as environment variables - 
 
 # token <- "activityinfotoken.."
 # print(Sys.setenv(ACTIVITYINFOTOKEN = token))   
 # Sys.getenv("ACTIVITYINFOTOKEN")
 # rm(token)
}
\examples{
#prepare_qa_data(activityInfoTable= "cdn6y40lm87wi522", activityInfoPlan ="caqxz3lrqv8wig3", activityInfoCountry="cwpgxrfllcr4wv66")
data <- prepare_qa_data(activityInfoTable= "cdn6y40lm87wi522")
}
