% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_country.R
\name{report_country}
\alias{report_country}
\title{Generate a summary word report}
\usage{
report_country(year = 2023, activityInfoTable, folder = "Report")
}
\arguments{
\item{folder}{folder within your project where to put the generated report. 
Folder will be created if it does not exist}

\item{activityInfo}{activityInfo link to the db}

\item{resource}{file with budget resource}
}
\value{
nothing the file for the report is generated
}
\description{
Generate a summary word report
}
\examples{

# ProgQA::report_country(year = 2023, activityInfoTable= "cdn6y40lm87wi522",   folder = "dev/report")
}
