% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_resource.R
\name{result_resource}
\alias{result_resource}
\title{result_resource}
\usage{
result_resource(data, thisind, budget, thisoutcome, poptype = NULL)
}
\arguments{
\item{data}{table resulting from  prepare_qa_data()}

\item{thisind}{which indicator to chart}

\item{budget}{table  from results.unhcr.org 4.6.1_Budget_Download.xlsx}

\item{thisoutcome}{which outcome to chart}

\item{poptype}{poptype filter - default is null}
}
\value{
ggplot2
}
\description{
This display a charts to review jointly indicators and resource allocation
}
\examples{

data <- prepare_qa_data(activityInfoTable= "cdn6y40lm87wi522")
knitr::kable(data |>  
               dplyr::filter(means_verification_results_level == "Outcome" ) |>
               dplyr::arrange( means_verification_indicator_code) |>
               dplyr::select(  means_verification_outcome_area, 
                               means_verification_indicator) |>
               dplyr::distinct())
## budget data to be downloaded and save locally

budget <- readxl::read_excel( system.file("4.6.1_Budget_Download.xlsx", package = "ProgQA"),  skip = 1)|>  
#budget <- readxl::read_excel( here::here("data-raw", "4.6.1_Budget_Download.xlsx"),  skip = 1)|>
  janitor::clean_names() 

result_resource(data,
                budget, 
                thisind = "1.1 Proportion of refugees and asylum seekers registered on an individual basis.",
                thisoutcome = "OA1: Access/Doc")

result_resource(data, budget,
                        poptype = "Refugees and Asylum-seekers",
thisind = "13.3 Proportion of PoC (working age) who are unemployed.",  thisoutcome = "OA13: Livelihood")
}
