% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_qa_table.R
\name{summary_qa_table}
\alias{summary_qa_table}
\title{summary_qa_table}
\usage{
summary_qa_table(data)
}
\arguments{
\item{data}{table resulting from  prepare_qa_data()}
}
\description{
A summary table of main QA issues
}
\examples{
data <- prepare_qa_data(activityInfoTable= "cdn6y40lm87wi522")
summary_qa_table(data)
}
