% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incountry_gap.R
\name{incountry_gap}
\alias{incountry_gap}
\title{Generate a bullet chart for in country indicator gap analysis}
\usage{
incountry_gap(data, thisoperation_mco)
}
\arguments{
\item{data}{RBM QA dataset - expect a few pre-defined variables to works well..}

\item{thisoperation_mco}{which operation plan to chart}
}
\value{
list with a plot
}
\description{
This chart is designed to allow for quick gap analysis based on 
  all outcome indicators within the same country
}
\examples{
data <- prepare_qa_data(activityInfoTable= "cdn6y40lm87wi522")
incountry_gap(data, thisoperation_mco = "Brazil ABC" )
}
